fix: rename files and folders, remove singleon approach
