+++ src/script/conversation/ConversationVerificationStateHandler/MLS/index.ts
@@ -17,4 +17,4 @@
  *
  */
 
-export * from './proteusStateHandler';
+export * from './MLSStateHandler';
