+++ src/script/conversation/ConversationVerificationStateHandler/index.ts
@@ -17,4 +17,5 @@
  *
  */
 
-export * from './proteusStateHandler';
+export * from './Proteus';
+export * from './MLS';
